py/requirements.txt
# Discord Chat to XLSX Exporter - Requirements
# pip install -r requirements.txt

# Discord API library
discord.py>=2.3.0

# Data manipulation and analysis
pandas>=2.0.0

# Excel file support for pandas
openpyxl>=3.1.0

# Optional: For better date/time handling
python-dateutil>=2.8.0

# Optional: For progress bars (if you want to add them later)
tqdm>=4.65.0

# Optional: For configuration file support
pyyaml>=6.0
